refactor(codelenses): extract TestSourcesProvider, drop @Lazy self-injection
В `*CodeLensSupplier`-ах `@Lazy @Autowired self` использовался только чтобы
позвать `@Cacheable getTestSources(Path)` из базового класса через
Spring AOP-прокси. Под native-image этот паттерн ломается: `@Lazy`-инъекция
создаёт runtime CGLIB-прокси через `ContextAnnotationAutowireCandidateResolver.
buildLazyResolutionProxy`, который конфликтует с pre-generated CGLIB-стабом
от Spring AOT (ClassCast: `CglibAopProxy$SerializableNoOp` → `Dispatcher`
в `setCallbacks`).

Выношу кэширование в отдельный `TestSourcesProvider`:
- `@Component @CacheConfig(cacheNames = "testSources")` с `@Cacheable getTestSources(Path)`
  и `@CacheEvict allEntries=true evict()`.
- `AbstractRunTestsCodeLensSupplier` принимает `TestSourcesProvider` через конструктор,
  избавляется от `@Cacheable`/`@CacheConfig`/`getSelf()`. Обработчики событий
  явно дёргают `testSourcesProvider.evict()` (раньше это делал `@CacheEvict` no-op).
- `DebugTestCodeLensSupplier`/`RunTestCodeLensSupplier`/`RunAllTestsCodeLensSupplier`
  больше не держат self-поля и лишние импорты `@Lazy`/`@Autowired`/`Lombok.Getter`,
  прокидывают `TestSourcesProvider` в `super(...)`.
- `AbstractRunTestsCodeLensSupplierTest.TestConfig` обновлён под новый конструктор.

Кэш по-прежнему живёт в Spring (через нормальный AOP-прокси на отдельном бине,
а не self-injection-через-`@Lazy`). Юнит-тесты Debug/Run/RunAllTests/AbstractRunTests
по-прежнему зелёные.
