+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/codelenses/AbstractRunTestsCodeLensSupplier.java
@@ -26,28 +26,20 @@ import com.github._1c_syntax.bsl.languageserver.configuration.events.LanguageSer
 import com.github._1c_syntax.bsl.languageserver.context.DocumentContext;
 import com.github._1c_syntax.bsl.languageserver.context.FileType;
 import com.github._1c_syntax.bsl.languageserver.events.LanguageServerInitializeRequestReceivedEvent;
-import com.github._1c_syntax.utils.Absolute;
 import lombok.RequiredArgsConstructor;
 import org.eclipse.lsp4j.ClientInfo;
 import org.eclipse.lsp4j.InitializeParams;
-import org.jspecify.annotations.Nullable;
-import org.springframework.cache.annotation.CacheConfig;
-import org.springframework.cache.annotation.CacheEvict;
-import org.springframework.cache.annotation.Cacheable;
 import org.springframework.context.event.EventListener;
 
 import java.net.URI;
-import java.nio.file.Path;
 import java.util.Optional;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 @RequiredArgsConstructor
-@CacheConfig(cacheNames = "testSources")
 public abstract class AbstractRunTestsCodeLensSupplier<T extends CodeLensData>
   implements CodeLensSupplier<T> {
 
   protected final LanguageServerConfiguration configuration;
+  protected final TestSourcesProvider testSourcesProvider;
 
   private boolean clientIsSupported;
 
@@ -59,7 +51,6 @@ public abstract class AbstractRunTestsCodeLensSupplier<T extends CodeLensData>
    * @param event Событие
    */
   @EventListener
-  @CacheEvict(allEntries = true)
   public void handleEvent(LanguageServerInitializeRequestReceivedEvent event) {
     var clientName = Optional.of(event)
       .map(LanguageServerInitializeRequestReceivedEvent::getParams)
@@ -67,19 +58,17 @@ public abstract class AbstractRunTestsCodeLensSupplier<T extends CodeLensData>
       .map(ClientInfo::getName)
       .orElse("");
     clientIsSupported = "Visual Studio Code".equals(clientName);
+    testSourcesProvider.evict();
   }
 
   /**
    * Обработчик события {@link LanguageServerConfigurationChangedEvent}.
    * <p>
    * Сбрасывает кеш при изменении конфигурации.
-   *
-   * @param event Событие
    */
-  @EventListener
-  @CacheEvict(allEntries = true)
-  public void handleLanguageServerConfigurationChange(LanguageServerConfigurationChangedEvent event) {
-    // No-op. Служит для сброса кеша при изменении конфигурации
+  @EventListener(LanguageServerConfigurationChangedEvent.class)
+  public void handleLanguageServerConfigurationChange() {
+    testSourcesProvider.evict();
   }
 
   /**
@@ -88,39 +77,11 @@ public abstract class AbstractRunTestsCodeLensSupplier<T extends CodeLensData>
   @Override
   public boolean isApplicable(DocumentContext documentContext) {
     var uri = documentContext.getUri();
-    var testSources = getSelf().getTestSources(documentContext.getServerContext().getConfigurationRoot());
+    var testSources = testSourcesProvider.getTestSources(documentContext.getServerContext().getConfigurationRoot());
 
     return clientIsSupported
       && documentContext.getFileType() == FileType.OS
       && testSources.stream().anyMatch(testSource -> isInside(uri, testSource));
-  }
-
-  /**
-   * Получить self-injected экземпляр себя для работы механизмов кэширования.
-   *
-   * @return Управляемый Spring'ом экземпляр себя
-   */
-  protected abstract AbstractRunTestsCodeLensSupplier<T> getSelf();
-
-  /**
-   * Получить список каталогов с тестами с учетом корня рабочей области.
-   * <p>
-   * public для работы @Cachable.
-   *
-   * @param configurationRoot Корень конфигурации
-   * @return Список исходных файлов тестов
-   */
-  @Cacheable
-  public Set<URI> getTestSources(@Nullable Path configurationRoot) {
-    var configurationRootString = Optional.ofNullable(configurationRoot)
-      .map(Path::toString)
-      .orElse("");
-
-    return configuration.getCodeLensOptions().getTestRunnerAdapterOptions().getTestSources()
-      .stream()
-      .map(testDir -> Path.of(configurationRootString, testDir))
-      .map(path -> Absolute.path(path).toUri())
-      .collect(Collectors.toSet());
   }
 
   private static boolean isInside(URI childURI, URI parentURI) {
