+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/codelenses/RunTestCodeLensSupplier.java
@@ -28,14 +28,11 @@ import com.github._1c_syntax.bsl.languageserver.context.FileType;
 import com.github._1c_syntax.bsl.languageserver.context.symbol.MethodSymbol;
 import com.github._1c_syntax.bsl.languageserver.utils.Resources;
 import lombok.EqualsAndHashCode;
-import lombok.Getter;
 import lombok.ToString;
 import lombok.Value;
 import lombok.extern.slf4j.Slf4j;
 import org.eclipse.lsp4j.CodeLens;
 import org.eclipse.lsp4j.Command;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Lazy;
 import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
 
@@ -61,19 +58,13 @@ public class RunTestCodeLensSupplier
   private final TestRunnerAdapter testRunnerAdapter;
   private final Resources resources;
 
-  // Self-injection для работы кэша в базовом классе.
-  @Autowired
-  @Lazy
-  @Getter
-  @SuppressWarnings("NullAway.Init")
-  private RunTestCodeLensSupplier self;
-
   public RunTestCodeLensSupplier(
     LanguageServerConfiguration configuration,
+    TestSourcesProvider testSourcesProvider,
     TestRunnerAdapter testRunnerAdapter,
     Resources resources
   ) {
-    super(configuration);
+    super(configuration, testSourcesProvider);
     this.testRunnerAdapter = testRunnerAdapter;
     this.resources = resources;
   }
